fix(nextjs-starter): tsc-esm and moving files
