+++ .gitignore
@@ -43,3 +43,4 @@ solana-signin-firebase-adminsdk-xijbn-40f35347e4.json
 /lib/**
 lerna-debug.log
 **/*/lib/**
+packages/examples/nextjs-starter/service-account.json
