+++ package.json
@@ -53,7 +53,9 @@
     "lerna": "^4.0.0",
     "next-transpile-modules": "^8.0.0",
     "shx": "^0.3.3",
-    "typescript": "^4.4.4"
+    "typescript": "^4.4.4",
+    "tsc-esm": "^1.0.4",
+    "typescript-esm": "^2.0.0"
   },
   "resolutions": {
     "@types/react": "^17.0.24"
