+++ packages/examples/nextjs-starter/utils/firebase.ts
@@ -1,15 +1,18 @@
-import { cert, getApps, initializeApp } from "firebase-admin/app";
-import { FIRESTORE_CLIENT_EMAIL, FIRESTORE_PRIVATE_KEY } from "../keys";
+
+import { cert, getApps, initializeApp, credential } from "firebase-admin/app";
+import * as serviceAccount from "../service-account.json"
 
 let firebaseApp = getApps()[0];
 
-if (!firebaseApp)
-  firebaseApp = initializeApp({
-    credential: cert({
-      projectId: "solana-signin",
-      privateKey: FIRESTORE_PRIVATE_KEY,
-      clientEmail: FIRESTORE_CLIENT_EMAIL,
-    }),
-  });
 
-export default firebaseApp;
+function getFirebaseApp(){
+if (!firebaseApp) {
+  firebaseApp = initializeApp({credential:  cert(serviceAccount)});
+
+}
+  return firebaseApp
+
+}
+
+
+export default getFirebaseApp;
