+++ packages/ui/react-ui/package.json
@@ -7,7 +7,7 @@
   "type": "module",
   "sideEffects": false,
   "main": "lib/cjs/index.js",
-  "module": "lib/esm/index.js",
+  "module": "lib/esm/index.mjs",
   "types": "lib/esm/index.d.ts",
   "files": [
     "lib",
@@ -15,7 +15,7 @@
     "LICENSE"
   ],
   "exports": {
-    "import": "./lib/esm/index.js",
+    "import": "./lib/esm/index.mjs",
     "require": "./lib/cjs/index.js"
   },
   "publishConfig": {
@@ -23,7 +23,7 @@
   },
   "scripts": {
     "clean": "shx rm -rf lib/*",
-    "build": "yarn clean && tsc -p tsconfig.json && tsc -p tsconfig-cjs.json",
+    "build": "yarn clean && tsc -p tsconfig.json && tsc-esm -p tsconfig.json && tsc -p tsconfig-cjs.json",
     "postbuild": "echo '{\"type\":\"commonjs\"}' | npx json > lib/cjs/package.json && echo '{\"type\":\"module\"}' | npx json > lib/esm/package.json"
   },
   "dependencies": {
